projects/ai-newsletter-crew/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
crewai==0.86.0
python-dotenv==1.0.0
duckduckgo-search==5.3.0
eventlet==0.33.3
